+++ lib/flutter_vscode.dart
@@ -1,6 +1,0 @@
-
-export 'annotations.dart';
-export 'src/vscode.dart';
-export 'src/vscode_controller_base.dart';
-export 'src/vscode_webview_helper.dart';
-export 'src/webview_bridge.dart';