+++ lib/src/webview_bridge.dart
@@ -1,1 +1,0 @@
-export 'webview_bridge_stub.dart' if (dart.library.html) 'webview_bridge_web.dart';